+++ examples/basic-harmony/src/app-worker.ts
@@ -1,0 +1,5 @@
+export default {
+  onCreate() {
+    console.info('app created');
+  }
+}