+++ examples/with-atomic-css/src/components/BuggyCounter/index.tsx
@@ -1,0 +1,27 @@
+import { createElement, Component } from 'rax';
+import Text from 'rax-text';
+
+class BuggyCounter extends Component {
+  constructor(props) {
+    super(props);
+    this.state = { counter: 0 };
+    this.handleClick = this.handleClick.bind(this);
+  }
+
+  handleClick() {
+    this.setState(({ counter }) => ({
+      counter: counter + 1,
+    }));
+  }
+
+  render() {
+    if (this.state.counter === 5) {
+      // Simulate a JS error
+      throw new Error('I crashed!');
+    }
+    return <Text onClick={this.handleClick}>{this.state.counter}</Text>;
+  }
+}
+
+
+export default BuggyCounter;