+++ examples/with-rax-mpa/src/pages/Home/index.tsx
@@ -7,7 +7,7 @@ export default function Home(props) {
   return (
     <div className="home">
       <Logo/>
-      <text className="title">Welcome to Your Rax App</text>
+      <text className="title">Welcome to Your Rax App123</text>
       <text className="info">More information about Rax</text>
     </div>
   );