+++ packages/plugin-ssr/src/index.ts
@@ -99,7 +99,7 @@ export default function (api) {
       const minifedHtml = minify(html, { collapseWhitespace: true, quoteCharacter: "'" });
       const newBundle = bundle
         .replace(/__RAX_APP_SERVER_HTML_TEMPLATE__/, minifedHtml)
-        .replace(new RegExp('__CHUNK_INFO__'), JSON.stringify(chunkInfo));
+        .replace(new RegExp('__CHUNK_INFO__'), encodeURIComponent(JSON.stringify(chunkInfo)));
       fs.writeFileSync(serverFilePath, newBundle, 'utf-8');
     });
   });