+++ packages/plugin-ssr/src/ssr/EntryLoader.ts
@@ -1,9 +1,10 @@
 import { formatPath } from '@builder/app-helpers';
-import * as qs from 'qs';
+import { parse } from 'query-string';
 import * as path from 'path';
-import { ILoaderQuery } from '../types';
+import { IFormattedLoaderQuery, ILoaderQuery } from '../types';
 import addCustomRenderComponentToHTML from './addCustomRenderComponentToHTML';
 import addBuiltInRenderComponentToHTML from './addBuiltInRenderComponentToHTML';
+import formatEntryLoaderQuery from '../utils/formatEntryLoaderQuery';
 
 function addExport(code) {
   return `${code}
@@ -18,43 +19,48 @@ function addExport(code) {
 function addDefineInitialPage() {
   return `
   const pathname = req.path;
-  let Page = appConfig.renderComponent;
-  if (!Page) {
+  let Page;
+  if (enableRouter) {
     const route = staticConfig.routes.find(({ path }) => path === pathname);
     Page = route.component;
+  } else {
+    Page = appConfig.app && appConfig.app.renderComponent;
   }
   `;
 }
 
 /**
  * Global variables:
- * utils: Generator
+ * utils: Generator, parseUrl
  * rax render: createElement/renderer
  * Component: Page/Document
- * generated in .rax: appConfig/staticConfig/createBaseApp/emitLifeCycles
+ * generated in .rax: appConfig/staticConfig/createBaseApp/emitLifeCycles/setHistory/enableRouter
  * MPA: pageConfig
  */
 export default function () {
-  const query = (qs.parse(this.query.substr(1)) as unknown) as ILoaderQuery;
+  const query = parse(this.query) as unknown as ILoaderQuery;
+  const formattedQuery: IFormattedLoaderQuery = formatEntryLoaderQuery(query);
   const corePath = path.join(query.tempPath, 'core');
-  query.needInjectStyle = query.needInjectStyle === 'true';
-  query.updateDataInClient = query.updateDataInClient === 'true';
+
   let code = `
     import Generator from '@builder/html-generator';
     import { createElement } from 'rax';
     import renderer from 'rax-server-renderer';
+    import * as parseUrl from 'parse-url';
     import { getAppConfig } from '${formatPath(path.join(corePath, 'appConfig'))}';
     import { emitLifeCycles } from '${formatPath(path.join(corePath, 'publicAPI'))}';
+    import { setHistory } from '${formatPath(path.join(corePath, 'routerAPI'))}';
     import '${formatPath(this.resourcePath)}';
-    import app from '${formatPath(query.runAppPath)}';
+    import app from '${formatPath(formattedQuery.runAppPath)}';
 
-    const { createBaseApp, staticConfig, pageConfig, TabBar } = app;
+    const { createBaseApp, staticConfig, pageConfig, TabBar, enableRouter } = app;
 
     const escapeLookup = {
       '<': '\\u003c',
       '>': '\\u003e',
       '/': '\\u002f',
     };
+
     const escapeRegex = /[<>\/]/g;
     function stripXSS(str) {
       return str.replace(escapeRegex, (match) => escapeLookup[match]);
@@ -73,19 +79,39 @@ export default function () {
       throw new Error(message);
     }
 
-    ${query.documentPath ? addCustomRenderComponentToHTML(query) : addBuiltInRenderComponentToHTML(query)}
+    function getInitialContext(ctx) {
+      const { req, res } = ctx;
+      const { hash, search, query } = parseUrl(req.url);
+      const pathname = req.path;
+      const location = {
+        pathname,
+        search,
+        state: null,
+        hash,
+      };
+
+      return {
+        req,
+        res,
+        pathname,
+        query,
+        location,
+      };
+    }
+
+    ${formattedQuery.documentPath ? addCustomRenderComponentToHTML(formattedQuery) : addBuiltInRenderComponentToHTML(formattedQuery)}
 
     async function renderToHTML(req, res, options = {}) {
       const { initialData, htmlTemplate, chunkInfo } = options;
       ${addDefineInitialPage()}
-      const html = await renderComponentToHTML(Page, { req, res }, initialData, htmlTemplate, chunkInfo);
+      const html = await renderComponentToHTML(Page, getInitialContext({ req, res }), initialData, htmlTemplate, chunkInfo);
       return html;
     }
 
     async function renderToHTMLWithContext(ctx, options = {}) {
       const { initialData, htmlTemplate, chunkInfo } = options;
       ${addDefineInitialPage()}
-      const html = await renderComponentToHTML(Page, ctx, initialData, htmlTemplate, chunkInfo);
+      const html = await renderComponentToHTML(Page, getInitialContext(ctx), initialData, htmlTemplate, chunkInfo);
       return html;
     }
 
@@ -95,7 +121,7 @@ export default function () {
       ${addDefineInitialPage()}
       let html;
       try {
-        html = await renderComponentToHTML(Page, ctx, initialData);
+        html = await renderComponentToHTML(Page, getInitialContext(ctx), initialData);
       } catch (e) {
         html = htmlTemplate;
         console.error(e);
@@ -125,8 +151,8 @@ export default function () {
     }
   `;
 
-  if (query.documentPath) {
-    code = `import Document from '${formatPath(query.documentPath)}';
+  if (formattedQuery.documentPath) {
+    code = `import Document from '${formatPath(formattedQuery.documentPath)}';
     ${code}`;
   }