+++ packages/webpack-dev-server-v4/index.js
@@ -1,3 +1,0 @@
-const webpackDevServer = require('webpack-dev-server/lib/Server');
-
-module.exports = webpackDevServer;