+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+## 3.7.4
+
+- Chore: add node version validate
+- Fix: platform-loader error with no match platform
+- Fix: inject script or stylesheet by assets
+
 ## 3.7.3
 
 - Chore: lock webpackbar version for avoid minimum node version error